# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API para gestión de homologaciones y dispositivos",
        "title": "LBTech API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@lbtech.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/device-tracking": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Obtiene la lista de todos los trackings de dispositivos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device Tracking"
                ],
                "summary": "Listar trackings de dispositivos",
                "responses": {
                    "200": {
                        "description": "Lista de trackings",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.DeviceTracking"
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Agrega un nuevo log a un tracking existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device Tracking"
                ],
                "summary": "Agregar log de tracking",
                "parameters": [
                    {
                        "description": "Datos del log",
                        "name": "trackingLog",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.TrackingLogMultiple"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Log agregado exitosamente"
                    },
                    "400": {
                        "description": "Datos inválidos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflicto",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Crea un nuevo registro de tracking para un dispositivo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device Tracking"
                ],
                "summary": "Crear nuevo tracking de dispositivo",
                "parameters": [
                    {
                        "description": "Datos del tracking",
                        "name": "deviceTracking",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.DeviceTracking"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Tracking creado exitosamente"
                    },
                    "400": {
                        "description": "Datos inválidos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Tracking duplicado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/device-tracking/advanced-search": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Realiza una búsqueda avanzada de trackings con filtros",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device Tracking"
                ],
                "summary": "Búsqueda avanzada de trackings",
                "parameters": [
                    {
                        "description": "Opciones de búsqueda",
                        "name": "searchOption",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.SearchOption"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultados de la búsqueda",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.DeviceTracking"
                            }
                        }
                    },
                    "400": {
                        "description": "Parámetros de búsqueda inválidos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/device-tracking/search-options": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Obtiene las opciones disponibles para la búsqueda avanzada",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device Tracking"
                ],
                "summary": "Obtener opciones de búsqueda avanzada",
                "responses": {
                    "200": {
                        "description": "Opciones de búsqueda",
                        "schema": {
                            "$ref": "#/definitions/responses.SearchOption"
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/device-tracking/{id}": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Actualiza un tracking de dispositivo existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device Tracking"
                ],
                "summary": "Actualizar tracking",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del tracking",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos actualizados",
                        "name": "deviceTracking",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.DeviceTrackingExpanded"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tracking actualizado exitosamente"
                    },
                    "400": {
                        "description": "Datos inválidos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Tracking no encontrado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflicto con datos existentes",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Elimina un tracking de dispositivo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device Tracking"
                ],
                "summary": "Eliminar tracking",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del tracking",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tracking eliminado exitosamente"
                    },
                    "401": {
                        "description": "No autorizado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Tracking no encontrado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/export/device-tracking": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Exporta los trackings de dispositivos a un archivo Excel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "Device Tracking"
                ],
                "summary": "Exportar trackings a Excel",
                "parameters": [
                    {
                        "description": "Opciones de búsqueda/filtro",
                        "name": "searchOption",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.SearchOption"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Archivo Excel",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Parámetros inválidos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.Company": {
            "type": "object",
            "required": [
                "address",
                "email",
                "name"
            ],
            "properties": {
                "_id": {
                    "type": "string"
                },
                "address": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "logo_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "request.DeviceTracking": {
            "type": "object",
            "required": [
                "company",
                "device"
            ],
            "properties": {
                "company": {
                    "type": "string"
                },
                "device": {
                    "type": "string"
                },
                "imei": {
                    "type": "string"
                },
                "imeis": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "tracking_log": {
                    "$ref": "#/definitions/request.TrackingLog"
                },
                "tracking_logs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.TrackingLog"
                    }
                }
            }
        },
        "request.DeviceTrackingExpanded": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "company": {
                    "$ref": "#/definitions/request.Company"
                },
                "device": {
                    "$ref": "#/definitions/request.DeviceResume"
                },
                "imei": {
                    "type": "string"
                },
                "tracking_logs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.TrackingLog"
                    }
                }
            }
        },
...
        "request.SearchOption": {
            "type": "object",
            "properties": {
                "brands": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "commercial_models": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "countries": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "locations": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "request.TrackingLogMultiple": {
            "type": "object",
            "required": [
                "device_trackings",
                "tracking_log"
            ],
            "properties": {
                "device_trackings": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "tracking_log": {
                    "$ref": "#/definitions/request.TrackingLog"
                }
            }
        },
...
        "responses.DeviceTracking": {
            "type": "object",
            "properties": {
                "company": {
                    "type": "string"
                },
                "device": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "imei": {
                    "type": "string"
                },
                "trackingLogs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.TrackingLog"
                    }
                }
            }
        },
...
        "responses.SearchOption": {
            "type": "object",
            "properties": {
                "brands": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "commercial_models": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "countries": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "locations": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "responses.TrackingLog": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "country": {
                    "$ref": "#/definitions/responses.Country"
                },
                "document_url": {
                    "type": "string"
                },
                "external_responsible": {
                    "type": "string"
                },
                "internal_responsible": {
                    "$ref": "#/definitions/responses.User"
                },
                "location": {
                    "$ref": "#/definitions/responses.Location"
                },
                "person": {
                    "$ref": "#/definitions/responses.Person"
                },
                "tracking_date": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}